make utilities/CMakeLists.txt respect RF24_NO_IRQ
This is needed to update the pyRF24 pkg to use latest master.

This does not affect RF24 lib builds because
RF24_NO_IRQ is respected in the root CMakeLists.txt.
But, the pyRF24 pkg doesn't actually use the root CMakeLists.txt.
Instead, the pyRF24 pkg only uses utility/CMakeLists.txt.